INTG-265 moved serial num formatter to general file
