+++ src/utils/common/helpers.ts
@@ -1,0 +1,14 @@
+// eslint-disable-next-line import/prefer-default-export
+export function serialNumberFormatter(currentSerialNumber: string): string {
+  let formatSerialNumber = currentSerialNumber.replaceAll('-', '');
+
+  if (formatSerialNumber.length > 3 && formatSerialNumber.length <= 6)
+    formatSerialNumber = `${formatSerialNumber.slice(0, 3)}-${formatSerialNumber.slice(3)}`;
+  else if (currentSerialNumber.length > 6)
+    formatSerialNumber = `${formatSerialNumber.slice(0, 3)}-${formatSerialNumber.slice(
+      3,
+      6
+    )}-${formatSerialNumber.slice(6)}`;
+
+  return formatSerialNumber;
+}
