+++ src/utils/common/pair/pairFormHelpers.test.ts
@@ -2,7 +2,7 @@ import { cleanup } from '@testing-library/react';
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore
 import replaceAllInserter from 'string.prototype.replaceall';
-import { isHttps, serialNumberFormatter } from './pairFormHelpers';
+import { isHttps } from './pairFormHelpers';
 
 replaceAllInserter.shim();
 
@@ -15,16 +15,5 @@ describe('Test pairFormHelpers()', () => {
 
     // Assert
     expect(result).toBeFalsy();
-  });
-
-  test('should return serial number with dash when running serialNumberFormatter()', () => {
-    // Arrange
-    const inputSerialNumber = '123456789';
-
-    // Act
-    const result = serialNumberFormatter(inputSerialNumber);
-
-    // Assert
-    expect(result).toBe('123-456-789');
   });
 });
